fix: handle anonymous migrations below v8.37

+++ database/migrations/2023_09_15_000001_create_user_columns.php
@@ -4,6 +4,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
+use Symfony\Component\Console\Output\ConsoleOutput;
+
 return new class extends Migration
 {
     /**
@@ -11,9 +13,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('sub')->nullable()->index();
-        });
+        if (version_compare(app()->version(), '8.37', '>=')) {
+            if (!Schema::hasColumn('users', 'sub')) {
+                Schema::table('users', function (Blueprint $table) {
+                    $table->string('sub')->nullable()->index();
+                });                
+            } else {
+                $output = new ConsoleOutput();
+                $output->writeln('The users table has the sub column. Skipping adding sub column.');
+            } //End if
+        } else {
+            throw new \Exception('Laravel version is not supported. Works only with Laravel 8.37 or higher.');
+        } //End if
     }
 
     /**
@@ -21,10 +32,15 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn([
-                'sub',
-            ]);
-        });
+        if (Schema::hasColumn('users', 'sub')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn([
+                    'sub',
+                ]);
+            });
+        } else {
+            $output = new ConsoleOutput();
+            $output->writeln('The users table does not have the sub column. Skipping dropping sub column.');
+        } //End if
     }
 };
